x2t - binary - native convert locked canvas - bix bug #28246

+++ PdfReader/Src/StringExt.h
@@ -73,11 +73,11 @@ namespace PdfReader
 
 	static std::wstring* AStringToPWString(const char* sString)
 	{
-		return new std::wstring(NSStringExt::CConverter::GetUnicodeFromSingleByteString((unsigned char*)sString, strlen(sString)));
+		return new std::wstring(NSStringExt::CConverter::GetUnicodeFromSingleByteString((unsigned char*)sString, (long)strlen(sString)));
 	}
 	static std::wstring AStringToWString(const char* sString)
 	{
-		return std::wstring(NSStringExt::CConverter::GetUnicodeFromSingleByteString((unsigned char*)sString, strlen(sString)));
+		return std::wstring(NSStringExt::CConverter::GetUnicodeFromSingleByteString((unsigned char*)sString, (long)strlen(sString)));
 	}
